+++ x/ecocredit/basket/types/v1/state_basket_class.go
@@ -1,0 +1,21 @@
+package v1
+
+import (
+	"cosmossdk.io/errors"
+
+	"github.com/regen-network/regen-ledger/x/ecocredit"
+	"github.com/regen-network/regen-ledger/x/ecocredit/core"
+)
+
+// Validate performs basic validation of the BasketClass state type
+func (m *BasketClass) Validate() error {
+	if m.BasketId == 0 {
+		return ecocredit.ErrParseFailure.Wrapf("basket id cannot be zero")
+	}
+
+	if err := core.ValidateClassID(m.ClassId); err != nil {
+		return errors.Wrap(err, "class id") // returns parse error
+	}
+
+	return nil
+}