+++ x/ecocredit/client/tx.go
@@ -15,9 +15,9 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/regen-network/regen-ledger/types"
-	basketcli "github.com/regen-network/regen-ledger/x/ecocredit/client/basket"
-	"github.com/regen-network/regen-ledger/x/ecocredit/core"
+	basketclient "github.com/regen-network/regen-ledger/x/ecocredit/basket/client"
 	marketclient "github.com/regen-network/regen-ledger/x/ecocredit/marketplace/client"
+	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 )
 
 const (
@@ -51,9 +51,9 @@ func TxCmd(name string) *cobra.Command {
 		TxUpdateClassAdminCmd(),
 		TxUpdateProjectAdminCmd(),
 		TxUpdateProjectMetadataCmd(),
-		basketcli.TxCreateBasketCmd(),
-		basketcli.TxPutInBasketCmd(),
-		basketcli.TxTakeFromBasketCmd(),
+		basketclient.TxCreateBasketCmd(),
+		basketclient.TxPutInBasketCmd(),
+		basketclient.TxTakeFromBasketCmd(),
 		marketclient.TxSellCmd(),
 		marketclient.TxUpdateSellOrdersCmd(),
 		marketclient.TxBuyDirectCmd(),