+++ x/ecocredit/server/expected_keepers.go
@@ -11,7 +11,7 @@ import (
 	basketapi "github.com/regen-network/regen-ledger/api/regen/ecocredit/basket/v1"
 	marketapi "github.com/regen-network/regen-ledger/api/regen/ecocredit/marketplace/v1"
 	api "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
-	"github.com/regen-network/regen-ledger/x/ecocredit/basket"
+	baskettypes "github.com/regen-network/regen-ledger/x/ecocredit/basket/types/v1"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 	markettypes "github.com/regen-network/regen-ledger/x/ecocredit/marketplace/types/v1"
 )
@@ -24,6 +24,6 @@ type Keeper interface {
 	RegisterInvariants(sdk.InvariantRegistry)
 	InitGenesis(ctx sdk.Context, cdc codec.JSONCodec, data json.RawMessage) ([]types.ValidatorUpdate, error)
 	ExportGenesis(ctx sdk.Context, cdc codec.JSONCodec) (json.RawMessage, error)
-	QueryServers() (core.QueryServer, basket.QueryServer, markettypes.QueryServer)
+	QueryServers() (core.QueryServer, baskettypes.QueryServer, markettypes.QueryServer)
 	GetStateStores() (api.StateStore, basketapi.StateStore, marketapi.StateStore)
 }