+++ x/ecocredit/simulation/operations.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/regen-network/regen-ledger/types/math"
 	"github.com/regen-network/regen-ledger/x/ecocredit"
-	"github.com/regen-network/regen-ledger/x/ecocredit/basket"
+	baskettypes "github.com/regen-network/regen-ledger/x/ecocredit/basket/types/v1"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 	marketsims "github.com/regen-network/regen-ledger/x/ecocredit/marketplace/simulation"
 	markettypes "github.com/regen-network/regen-ledger/x/ecocredit/marketplace/types/v1"
@@ -81,7 +81,7 @@ var (
 func WeightedOperations(
 	appParams simtypes.AppParams, cdc codec.JSONCodec,
 	ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient core.QueryServer, basketQryClient basket.QueryServer,
+	qryClient core.QueryServer, basketQryClient baskettypes.QueryServer,
 	mktQryClient markettypes.QueryServer) simulation.WeightedOperations {
 
 	var (